+++ FileFormats/HystemExtran_REG.vb
@@ -265,7 +265,7 @@ Namespace Fileformats
             FiStr.Close()
 
             'store time series
-            Me.FileTimeSeries.Add(sInfo.Index, ts)
+            Me.TimeSeries.Add(sInfo.Index, ts)
 
         End Sub